--- src/main/java/com/team9/PrimaryController.java
@@ -1,12 +1,0 @@
-package com.team9;
-
-import java.io.IOException;
-import javafx.fxml.FXML;
-
-public class PrimaryController {
-
-    @FXML
-    private void switchToSecondary() throws IOException {
-        App.setRoot("secondary");
-    }
-}
