+++ src/main/resources/com/team9/calculatorView.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.control.Button?>
 <?import javafx.geometry.Insets?>
 
-<VBox alignment="CENTER" spacing="20.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.team9.PrimaryController">
+<VBox alignment="CENTER" spacing="20.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.team9.CalculatorController">
     <children>
         <Label text="Primary View" />
         <Button fx:id="primaryButton" text="Switch to Secondary View" onAction="#switchToSecondary"/>
