+++ src/tests/java/com/team9/CalculatorTest.java
@@ -8,41 +8,41 @@ public class CalculatorTest {
 
     @Test
     void testAddition() {
-        assertEquals(2, CalculatorController.add(1, 1));
+        assertEquals(2, CalculatorInterface.add(1, 1));
     }
 
     @Test
     public void testSubtraction() {
-        assertEquals(5, CalculatorController.subtract(10, 5));
+        assertEquals(5, CalculatorInterface.subtract(10, 5));
     }
 
     @Test
     public void testMultiplication() {
-        assertEquals(21, CalculatorController.multiply(7, 3));
+        assertEquals(21, CalculatorInterface.multiply(7, 3));
     }
 
     @Test
     public void testDivision() {
-        assertEquals(6, CalculatorController.divide(42, 7));
+        assertEquals(6, CalculatorInterface.divide(42, 7));
     }
 
     @Test
     public void testZeroDivision() {
-        assertEquals(0, CalculatorController.divide(13, 0));
+        assertEquals(0, CalculatorInterface.divide(13, 0));
     }
 
     @Test
     public void testSquare() {
-        assertEquals(144, CalculatorController.square(12));
+        assertEquals(144, CalculatorInterface.square(12));
     }
 
     @Test
     public void testSquareRoot() {
-        assertEquals(9, CalculatorController.squareRoot(81));
+        assertEquals(9, CalculatorInterface.squareRoot(81));
     }
 
     @Test
     public void testSquareRootNegative() {
-        assertEquals(0, CalculatorController.squareRoot(-1));
+        assertEquals(0, CalculatorInterface.squareRoot(-1));
     }
 }
